LayoutTests/fast/dom/Window/Location/location-override-toString-on-proto-using-with.html

<html>
<head>
    <script src="../../../js/resources/js-test-pre.js"></script>
    <script src="resources/location-tests-functions.js"></script>
</head>
<body>
<script>
    with (window.location.__proto__) {
        var toString = function() { return "haxored"; }
    }

    var result = normalizeURL(String(window.location));
    var correctValue = normalizeURL(document.URL);
    shouldBe("result", "correctValue");
</script>
<script src="../../../js/resources/js-test-post.js"></script>
</body>
</html>
